refactor:新形式への対応のための整頓
