+++ app/domain/warnings/warning_message_builder.rb
@@ -2,7 +2,7 @@ module Warnings
 class WarningMessageBuilder
   def self.build(warning)
   # もし古い形式の構造など互換性のない形式のwarningが含まれていたらエラーを出す
-  unless warning[:concept] || warning[:pattern] || warning[:matches]
+  unless warning[:concept] || warning[:pattern] || warning[:matches] || warning[:field]
     raise ArgumentError, "Unsupported warning format: #{warning.inspect}"
   end
     concept_name = warning[:concept].to_s.camelize
