+++ app/forms/if_then_rule_form.rb
@@ -72,7 +72,7 @@ class IfThenRuleForm
     @warnings += Warnings::ThenOversizedActionChecker.check(then_action)
     @warnings += Warnings::ThenNegativeExpressionActionChecker.check(then_action)
     # 以下,他のレコードが関わる構成系
-    @warnings += ::IfConditionDuplicateChecker.check(
+    @warnings += Warnings::IfConditionDuplicateChecker.check(
       user: @current_user,
       if_condition: if_condition,
       exclude_id: @if_then_rule_of_model&.id
